requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django_environ==0.10.0